Alterando códigos da questão 1 e 4

+++ Lista 5/1.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int potencia(int x, int n)
+float potencia(int x, int n)
 {
     if (n == 0)
         return 1;
